+++ extensions/spaceship/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Spaceship Changelog
 
+## [] - {PR_MERGE_DATE}
+
+- In "DNS Records": 
+    1. View TTL
+    2. See Inactive Records
+
 ## [Initial Version] - 2025-02-10
 
 View Domains and View DNS Records.